The custom class-name-resolver enables to discriminate persistent objects from the selected ResultSet - eg. by using more column values to make a decision on which class should be instantiated from this DB row. This calculation can be of any complexity using the DB row from ResultSet. Current implementation only allows for discriminating on one string column which might be enough in many cases - but this custom mechanism enables any complexity for resolving persistent classes based on a DB row.
